Fix coordinate computation when new mask boundary segment is determined
The mask has to obey the projector transformation when it is drawn, but the
mask points themselves are stored without the project transformation.

When the user clicks near a mask line segment, it is split in two segments
and a new point is installed between the two ends of the segment. When the
line segment is determined, the cursor position must not have the project
transformation applied so that the calculated values are in the same
coordinate space as the mask points. But the projector transformation is
applied too early. Move it later.

This fixes cases where the wrong line segment was found provided that the
projector is moved away from the center.

The determination whether a mask point was clicked happens later when all
points involved have the projector transformation applied and is not
affected by this change.